Move base module from model to root
